updaes

+++ predict.py
@@ -166,9 +166,10 @@ def reload_model(rank, model_path, device):
     """Reload model to combat memory fragmentation"""
     torch.cuda.empty_cache()
     gc.collect()
-    new_model = AutoModelForSequenceClassification.from_pretrained(model_path).to(
-        device
-    )
+    new_model = AutoModelForSequenceClassification.from_pretrained(
+        model_path,
+        torch_dtype=torch.bfloat16,
+    ).to(device)
     new_model = DDP(new_model, device_ids=[rank], find_unused_parameters=False)
     torch.cuda.empty_cache()
     return new_model
@@ -410,9 +411,10 @@ def process_and_save_ddp(rank, cfg, world_size):
         )
 
     try:
-        model = AutoModelForSequenceClassification.from_pretrained(cfg.model_path).to(
-            device
-        )
+        model = AutoModelForSequenceClassification.from_pretrained(
+            cfg.model_path,
+            torch_dtype=torch.bfloat16,
+        ).to(device)
         model = DDP(model, device_ids=[rank], find_unused_parameters=False)
 
         with open(f"{cfg.model_path}/config.json", "r") as config_file:
